test: WIP llm evaluation
