+++ tests/fixtures/test_chat/1_how_many_pods/kubectl_get.txt.AUTOGENERATED
@@ -1,0 +1,6 @@
+{"toolset_name":"kubernetes/core","tool_name":"kubectl_get","match_params":{"kind":"pod","name":"grafana-69946c9bd6-kwjb6","namespace":"default"}}
+Command `kubectl get --show-labels -o wide pod grafana-69946c9bd6-kwjb6 -n default` failed with return code 1
+stdout:
+
+stderr:
+Error from server (NotFound): pods "grafana-69946c9bd6-kwjb6" not found
