+++ tests/fixtures/test_chat/1_how_many_pods/kubectl_top_pods.txt
@@ -1,0 +1,6 @@
+{"toolset_name":"kubernetes/live-metrics","tool_name":"kubectl_top_pods","match_params":{}}
+Command `kubectl top pods -A` failed with return code 1
+stdout:
+
+stderr:
+error: Metrics API not available
